Those changes are made to emphasize that the main version of the programm is the "advanced version" ` since is more user frendly. The default single use script is now placed in to the "simple version" folder. Other than that some Readme.md changes to keep it up to date.
