--- advanced_version/requirements.txt
@@ -1,1 +1,0 @@
-selenium==4.20.0
