--- advanced_version/todo.txt
@@ -1,17 +1,0 @@
-Are you certain game enjoyer? Do you fell annoyed selecting game (same one) every time on startup of the program?
-I'll have solution for you!
-The new function I planned will check if config file (txt?) exists to automatically choose game on startup!
-It will be able to select the download mode too!
-The plan is to even add auto download - you create file with all parameters and don't need to type it into console!
-The program will do it for you!
-
-~That's the plan will see when or if I'm gonna implement that or any other functionality mentioned in this file, since for me - the current state is enough.
-
-+ better card download -> checking if card is downloaded if it's not or server went down for a second - retry (i = i), and only then downloading next card (i += 1)
-+ if server went down for a second and first card couldn't be downloaded - retry loading page at least 1 time
-+ packing app into exe
-+ selecting download location
-
-//before download check if card exists in download folder to skip already downloaded cards
-
-Maybe add config - "Don't close browser after fnish" - set default true -> user checks if cards are being downloaded if not manualy terminates spawned browser
