+++ config.ini
@@ -6,7 +6,7 @@ close_browser_after_download = 0
 close_program_after_download = 1
 
 [USER]
-browser_visibility = 0
+browser_visibility = 1
 card_download_interval = 5
 website_load_interval = 5
 close_browser_after_download = 1
